requirements.txt
# Core dependencies
Flask==2.3.3
Werkzeug==2.3.7

# Web scraping
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Data processing
pandas==2.1.3
numpy==1.26.2

# Database
SQLAlchemy==2.0.23

# Scheduling
APScheduler==3.10.4

# Email
smtplib2==0.2.1

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Utilities
python-dateutil==2.8.2
Pillow==10.1.0
openpyxl==3.1.2

# API and serialization
marshmallow==3.20.1
flask-cors==4.0.0

# Security
cryptography==41.0.8 